COMPLETE: Implement full coordinator pattern for all platforms
MAJOR ARCHITECTURAL COMPLETION:
- Re-enabled sensor and binary_sensor platforms in __init__.py
- Fixed duplicate return statement in async_setup_entry
- Completely rewrote sensor.py to use CoordinatorEntity pattern
- Completely rewrote binary_sensor.py to use CoordinatorEntity pattern
- Backed up old sensor files as sensor_old.py and binary_sensor_old.py

NEW SENSOR ARCHITECTURE:
✅ ArubaPortLinkStatusSensor - reads link status from coordinator cache
✅ ArubaPortActivitySensor - shows port activity with statistics
✅ ArubaSwitchFirmwareSensor - displays firmware version
✅ ArubaSwitchModelSensor - displays switch model
✅ ArubaSwitchConnectivitySensor - binary sensor for switch connectivity

KEY IMPROVEMENTS:
- All sensors inherit from CoordinatorEntity + SensorEntity
- Zero individual SSH calls - all read from coordinator cache
- Proper availability handling via coordinator.last_update_success
- Single SSH session architecture maintained across all platforms

PERFORMANCE BENEFITS:
- Eliminates concurrent SSH connection issues during setup
- Single coordinator manages data for 50+ entities efficiently
- Proper update synchronization across all platforms
- Fast, reliable integration setup without timeouts

This completes the coordinator pattern migration for the entire integration.
